pkg/pi: s/binary executable/binary/
Not all binaries analyzed by Pippi will be executables,
some will be shared libraries, object files, archives,
etc.

Also, it is conceivable that we may wish to use Pippi
for analyzing binary files containing data, not only
code. Especially if we implement a hexeditor with
quality of life improvements. (The current hexeditors
in Linux are quite poor.)