proto: add encoding and size fields to strings.proto
To be able to add support for extracting strings of
other encodings than ASCII and UTF-8, we should store
the original encoding of the strings extracted by
pi-strings.

Also, since the length in characters of an extracted
string may differ from the size in number of bytes
that the string contents took in the binary executable,
a size field is added.