cmd/pi-strings-go: add Go version of pi-strings
The intention with pi-strings-go is not to replace
the Rust version of pi-strings but to provide multiple
implementations of the same Protobuf API, such that
different implementations may be evaluated and
compared for performance, feature sets, etc.

Once we get a large enough set of non-trivial
compoents it will be fun to evaluate Rust vs. Go
in terms of performance for more intricate reversing
sessions where the garbage collector may kick in high
drive.