cmd/pi-disasm-new: implement simple disassembler
The design is of the disasm package mirrors that of the
image package in the Go standard library but for decoding
instructions of machine architecture formats instead of
image formats. Noteably, the disasm package uses a
RegisterFormat function to register new disassemblers.

A user of the disasm package includes the machine
architecture disassemblers relevant to them, by e.g.:

	import _ ".../disasm/x86"

To separate concerns regarding data/code distinction, the
disassembler simply parses a "database" containing a list
of instruction addresses.

A stand-alone-script has been included to produce a naive
database of instruction addresses using the objdump tool.
Note that objdump is known to include faults instruction
addresses, and as such, expect to see warning messages
printed to standard error when using the disassembler.