cmd/pippi: update to latest version of Wails
To make this update the following steps were taken.

1. generate a clean v0.17.0 wails project (`wails init`)
2. generate a clean v0.18.3 wails project (`wails init`)
3. check the diff between the versions (`diff -u -r wails0170/foo/frontend wails0183/foo/frontend`)
4. update main.ts as this was generated by the initial `wails init` script
5. update package.json to contain the new dependencies (note, copy from v0.18.3, and then adding our dependencies for Ace, etc)
6. remove package-lock.json and generate a new one
7. update go.mod and go.sum using `go get github.com/wailsapp/wails@v0.18.3`
8. ...
9. profit

Note, we may consider adding a note that `main.ts` along with the
polyfill.js file (and probably a few others) may be regenerated
only when running `wails init`, so it becomes a manual effort for
us to update these as Wails is updated. At the very least, we should
strive to make minimal changes to these files as we would like to be
able to simply copy them from a clean `wails init` frontend/src dir.

Fixes #49.