wailsapp/webview: silence web view warnings
`go install ./...` is not silence once more!

This is done by updating the go.mod with a
replacement directive until
wailsapp/webview#2 is merged and a new
release of wailsapp/wails is issued which
depends on the latest webview. At that point,
we should just get rid of the replacement
directive in the go.mod.